refactor: breaking nav.ts for better scoping
