+++ packages/client/composables/useNav.ts
@@ -1,12 +1,48 @@
-import type { ClicksContext, SlideRoute } from '@slidev/types'
-import type { ComputedRef } from 'vue'
-import { computed } from 'vue'
-import { configs } from '../env'
-import type { SlidevContextNav } from '../modules/context'
-import { getSlidePath, rawTree, tree, treeWithActiveStatuses } from '../logic/nav'
+import type { ClicksContext, SlideRoute, TocItem } from '@slidev/types'
+import type { ComputedRef, Ref, TransitionGroupProps } from 'vue'
+import { computed, ref, watch } from 'vue'
+import type { Router } from 'vue-router'
+import { getCurrentTransition } from '../logic/transition'
+import { getSlidePath } from '../logic/slides'
+import { useTocTree } from './useTocTree'
+import { skipTransition } from './hmr'
 import { slides } from '#slidev/slides'
 
-export function useNavBase(currentSlideRoute: ComputedRef<SlideRoute>, clicksContext: ComputedRef<ClicksContext>) {
+export interface SlidevContextNav {
+  slides: Ref<SlideRoute[]>
+  total: ComputedRef<number>
+  path: ComputedRef<string>
+  currentPage: ComputedRef<number>
+  currentSlideNo: ComputedRef<number>
+  currentSlideRoute: ComputedRef<SlideRoute>
+  currentTransition: ComputedRef<TransitionGroupProps | undefined>
+  currentLayout: ComputedRef<string>
+  nextRoute: ComputedRef<SlideRoute>
+  prevRoute: ComputedRef<SlideRoute>
+  clicksContext: ComputedRef<ClicksContext>
+  clicks: ComputedRef<number>
+  clicksTotal: ComputedRef<number>
+  hasNext: ComputedRef<boolean>
+  hasPrev: ComputedRef<boolean>
+  tocTree: ComputedRef<TocItem[]>
+  navDirection: Ref<number>
+  clicksDirection: Ref<number>
+  openInEditor: (url?: string) => Promise<boolean>
+  next: () => Promise<void>
+  prev: () => Promise<void>
+  nextSlide: () => Promise<void>
+  prevSlide: (lastClicks?: boolean) => Promise<void>
+  go: (page: number | string, clicks?: number) => Promise<void>
+  goFirst: () => Promise<void>
+  goLast: () => Promise<void>
+}
+
+export function useNavBase(
+  currentSlideRoute: ComputedRef<SlideRoute>,
+  clicksContext: ComputedRef<ClicksContext>,
+  queryClicks: Ref<number> = ref(0),
+  router?: Router,
+): SlidevContextNav {
   const total = computed(() => slides.value.length)
   const path = computed(() => getSlidePath(currentSlideRoute.value))
   const currentSlideNo = computed(() => currentSlideRoute.value.no)
@@ -18,17 +54,14 @@ export function useNavBase(currentSlideRoute: ComputedRef<SlideRoute>, clicksCon
   const hasNext = computed(() => currentSlideNo.value < slides.value.length || clicks.value < clicksTotal.value)
   const hasPrev = computed(() => currentSlideNo.value > 1 || clicks.value > 0)
 
-  async function downloadPDF() {
-    const { saveAs } = await import('file-saver')
-    saveAs(
-      typeof configs.download === 'string'
-        ? configs.download
-        : configs.exportFilename
-          ? `${configs.exportFilename}.pdf`
-          : `${import.meta.env.BASE_URL}slidev-exported.pdf`,
-      `${configs.title}.pdf`,
-    )
-  }
+  const navDirection = ref(0)
+  const clicksDirection = ref(0)
+
+  const currentTransition = computed(() => getCurrentTransition(navDirection.value, currentSlideRoute.value, prevRoute.value))
+
+  watch(currentSlideRoute, (next, prev) => {
+    navDirection.value = next.no - prev.no
+  })
 
   async function openInEditor(url?: string) {
     if (url == null) {
@@ -41,6 +74,57 @@ export function useNavBase(currentSlideRoute: ComputedRef<SlideRoute>, clicksCon
     return true
   }
 
+  const tocTree = useTocTree(slides)
+
+  async function next() {
+    clicksDirection.value = 1
+    if (clicksTotal.value <= queryClicks.value)
+      await nextSlide()
+    else
+      queryClicks.value += 1
+  }
+
+  async function prev() {
+    clicksDirection.value = -1
+    if (queryClicks.value <= 0)
+      await prevSlide()
+    else
+      queryClicks.value -= 1
+  }
+
+  async function nextSlide() {
+    clicksDirection.value = 1
+    if (currentSlideNo.value < slides.value.length)
+      await go(currentSlideNo.value + 1)
+  }
+
+  async function prevSlide(lastClicks = true) {
+    clicksDirection.value = -1
+    const next = Math.max(1, currentSlideNo.value - 1)
+    await go(next)
+    if (lastClicks && clicksTotal.value) {
+      router?.replace({
+        query: { ...router.currentRoute.value.query, clicks: clicksTotal.value },
+      })
+    }
+  }
+
+  function goFirst() {
+    return go(1)
+  }
+
+  function goLast() {
+    return go(total.value)
+  }
+
+  async function go(page: number | string, clicks?: number) {
+    skipTransition.value = false
+    await router?.push({
+      path: getSlidePath(page),
+      query: { ...router.currentRoute.value.query, clicks },
+    })
+  }
+
   return {
     slides,
     total,
@@ -49,6 +133,8 @@ export function useNavBase(currentSlideRoute: ComputedRef<SlideRoute>, clicksCon
     currentPage: currentSlideNo,
     currentSlideRoute,
     currentLayout,
+    currentTransition,
+    clicksDirection,
     nextRoute,
     prevRoute,
     clicksContext,
@@ -56,24 +142,32 @@ export function useNavBase(currentSlideRoute: ComputedRef<SlideRoute>, clicksCon
     clicksTotal,
     hasNext,
     hasPrev,
-    downloadPDF,
+    tocTree,
+    navDirection,
     openInEditor,
+    next,
+    prev,
+    go,
+    goLast,
+    goFirst,
+    nextSlide,
+    prevSlide,
   }
 }
 
-export function useFixedNav(currentSlideRoute: SlideRoute, clicksContext: ClicksContext): SlidevContextNav {
-  const noOp = async () => { }
+export function useFixedNav(
+  currentSlideRoute: SlideRoute,
+  clicksContext: ClicksContext,
+): SlidevContextNav {
+  const noop = async () => { }
   return {
     ...useNavBase(computed(() => currentSlideRoute), computed(() => clicksContext)),
-    next: noOp,
-    prev: noOp,
-    nextSlide: noOp,
-    prevSlide: noOp,
-    goFirst: noOp,
-    goLast: noOp,
-    go: noOp,
-    rawTree,
-    treeWithActiveStatuses,
-    tree,
+    next: noop,
+    prev: noop,
+    nextSlide: noop,
+    prevSlide: noop,
+    goFirst: noop,
+    goLast: noop,
+    go: noop,
   }
 }
